requirement.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
crispy-bootstrap4==2022.1
Django==4.2.1
django-heroku==0.3.1
django-bootstrap4==22.3
django-crispy-forms==2.0
django-utils==0.0.2
djangorestframework==3.14.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
mysqlclient==2.1.1
numpy==1.24.3
oauthlib==3.2.2
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
